feat(ui): add Download action to the files row menu
Files-only "Download" item that fetches the file by path and saves it.
Extract the viewer's blob/anchor download into a shared lib/download.ts
so the row menu and FileViewer share one implementation.
